pyproject.toml
# Copyright (c) 2022, Semiotic AI, Inc.  All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[tool.poetry]
name = "thegraph-gql-utils"
version = "0.0.0"  # Filled in by poetry-dynamic-versioning
authors = [
    "Alexis Asseman <alexis@semiotic.ai>"
]
description = "GraphQL language utilities for The Graph"
license = "Apache 2.0"
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: Apache Software License"
]
repository = "https://gitlab.com/semiotic-ai/thegraph-gql-utils"
packages = [
    { include = "gql_utils", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
graphql-core = "~3.2"
more-itertools = "~8.12"

[tool.poetry.dev-dependencies]
black = "^22"
isort = "~5.10"
pytest = "^7.1"
pylint = "^2.13.4"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"

[build-system]
requires = ["poetry_core>=1.2.0a2", "poetry-dynamic-versioning==1.0.0b1"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
